Updated Bard stats + unit names on maps
