+++ SolStandard/Content/TmxMaps/Chesslike_01.tmx
@@ -2215,14 +2215,14 @@
  <objectgroup id="6" name="Loot"/>
  <objectgroup id="7" name="Items"/>
  <objectgroup id="8" name="Units">
-  <object id="520" name="Dooticus IV" type="Unit" gid="480" x="320" y="448" width="32" height="32">
+  <object id="520" name="Dhalia" type="Unit" gid="480" x="320" y="448" width="32" height="32">
    <properties>
     <property name="Class" value="Bard"/>
     <property name="Commander" type="bool" value="true"/>
     <property name="Team" value="Blue"/>
    </properties>
   </object>
-  <object id="521" name="Calcius II" type="Unit" gid="485" x="288" y="224" width="32" height="32">
+  <object id="521" name="Crimson" type="Unit" gid="485" x="288" y="224" width="32" height="32">
    <properties>
     <property name="Class" value="Bard"/>
     <property name="Commander" type="bool" value="true"/>
