+++ SolStandard/Content/TmxMaps/Crossroads_01.tmx
@@ -5073,14 +5073,14 @@
     <property name="Team" value="Blue"/>
    </properties>
   </object>
-  <object id="520" name="Mandible I" type="Unit" gid="480" x="640" y="704" width="32" height="32">
+  <object id="520" name="Aqua" type="Unit" gid="480" x="640" y="704" width="32" height="32">
    <properties>
     <property name="Class" value="Bard"/>
     <property name="Commander" type="bool" value="true"/>
     <property name="Team" value="Blue"/>
    </properties>
   </object>
-  <object id="521" name="Calfius III" type="Unit" gid="485" x="640" y="288" width="32" height="32">
+  <object id="521" name="Amber" type="Unit" gid="485" x="640" y="288" width="32" height="32">
    <properties>
     <property name="Class" value="Bard"/>
     <property name="Commander" type="bool" value="true"/>
