+++ SolStandard/Content/TmxMaps/Debug_01.tmx
@@ -2340,13 +2340,13 @@
     <property name="Team" value="Blue"/>
    </properties>
   </object>
-  <object id="520" name="Fracturius I" type="Unit" gid="480" x="416" y="384" width="32" height="32">
+  <object id="520" name="Aqua" type="Unit" gid="480" x="416" y="384" width="32" height="32">
    <properties>
     <property name="Class" value="Bard"/>
     <property name="Team" value="Blue"/>
    </properties>
   </object>
-  <object id="521" name="Marrowon II" type="Unit" gid="485" x="416" y="224" width="32" height="32">
+  <object id="521" name="Rose" type="Unit" gid="485" x="416" y="224" width="32" height="32">
    <properties>
     <property name="Class" value="Bard"/>
     <property name="Commander" type="bool" value="true"/>
