+++ SolStandard/Content/TmxMaps/Debug_02.tmx
@@ -1949,13 +1949,13 @@
     <property name="Team" value="Blue"/>
    </properties>
   </object>
-  <object id="520" name="Fracturius I" type="Unit" gid="480" x="448" y="416" width="32" height="32">
+  <object id="520" name="Fran" type="Unit" gid="480" x="448" y="416" width="32" height="32">
    <properties>
     <property name="Class" value="Bard"/>
     <property name="Team" value="Blue"/>
    </properties>
   </object>
-  <object id="521" name="Marrowon II" type="Unit" gid="485" x="448" y="192" width="32" height="32">
+  <object id="521" name="Marilyn" type="Unit" gid="485" x="448" y="192" width="32" height="32">
    <properties>
     <property name="Class" value="Bard"/>
     <property name="Commander" type="bool" value="true"/>
