+++ SolStandard/Content/TmxMaps/Fortress_01.tmx
@@ -5566,14 +5566,14 @@
   </object>
  </objectgroup>
  <objectgroup id="8" name="Units">
-  <object id="520" name="Fracturius I" type="Unit" gid="480" x="1024" y="800" width="32" height="32">
+  <object id="520" name="Malady" type="Unit" gid="480" x="1024" y="800" width="32" height="32">
    <properties>
     <property name="Class" value="Bard"/>
     <property name="Commander" type="bool" value="true"/>
     <property name="Team" value="Blue"/>
    </properties>
   </object>
-  <object id="521" name="Marrowon II" type="Unit" gid="485" x="288" y="64" width="32" height="32">
+  <object id="521" name="Melody" type="Unit" gid="485" x="288" y="64" width="32" height="32">
    <properties>
     <property name="Class" value="Bard"/>
     <property name="Commander" type="bool" value="true"/>
