+++ SolStandard/Content/TmxMaps/Grassland_02.tmx
@@ -4052,7 +4052,7 @@
  </objectgroup>
  <objectgroup id="7" name="Items"/>
  <objectgroup id="8" name="Units">
-  <object id="811" name="Anklius" type="Unit" gid="480" x="576" y="192" width="32" height="32">
+  <object id="811" name="Aqua" type="Unit" gid="480" x="576" y="192" width="32" height="32">
    <properties>
     <property name="Class" value="Bard"/>
     <property name="Commander" type="bool" value="true"/>
@@ -4126,7 +4126,7 @@
     <property name="Class" value="Orc"/>
    </properties>
   </object>
-  <object id="827" name="Usurper" type="Unit" gid="485" x="608" y="736" width="32" height="32">
+  <object id="827" name="Amber" type="Unit" gid="485" x="608" y="736" width="32" height="32">
    <properties>
     <property name="Class" value="Bard"/>
     <property name="Commander" type="bool" value="true"/>
