+++ SolStandard/Content/TmxMaps/Scotia_Hill_Redux_01.tmx
@@ -10458,14 +10458,14 @@
     <property name="Team" value="Blue"/>
    </properties>
   </object>
-  <object id="520" name="King Krown" type="Unit" gid="480" x="576" y="704" width="32" height="32">
+  <object id="520" name="Cameron" type="Unit" gid="480" x="576" y="704" width="32" height="32">
    <properties>
     <property name="Class" value="Bard"/>
     <property name="Commander" type="bool" value="true"/>
     <property name="Team" value="Blue"/>
    </properties>
   </object>
-  <object id="521" name="Lord Ligament" type="Unit" gid="485" x="1024" y="1408" width="32" height="32">
+  <object id="521" name="Lisa" type="Unit" gid="485" x="1024" y="1408" width="32" height="32">
    <properties>
     <property name="Class" value="Bard"/>
     <property name="Commander" type="bool" value="true"/>
