+++ SolStandard/Content/TmxMaps/Village_01.tmx
@@ -7617,14 +7617,14 @@
     <property name="Team" value="Blue"/>
    </properties>
   </object>
-  <object id="520" name="Mandible I" type="Unit" gid="480" x="928" y="1088" width="32" height="32">
+  <object id="520" name="Malady" type="Unit" gid="480" x="928" y="1088" width="32" height="32">
    <properties>
     <property name="Class" value="Bard"/>
     <property name="Commander" type="bool" value="true"/>
     <property name="Team" value="Blue"/>
    </properties>
   </object>
-  <object id="521" name="Calfius III" type="Unit" gid="485" x="544" y="512" width="32" height="32">
+  <object id="521" name="Callie" type="Unit" gid="485" x="544" y="512" width="32" height="32">
    <properties>
     <property name="Class" value="Bard"/>
     <property name="Commander" type="bool" value="true"/>
